+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG for 2.x
 * [New] Package manager with the following new commands: "new:theme", "add:plugin", "remove:plugin" and "update:plugin". Additionaly, "new:site" command has been updated for scaffolding new sites using themes. See the new feature #96.
 * [New] Allow default attributes on "pages" collection. See new feature #93.
 * [New] New options for applying "avoid_renderizer" feature to type of files and files belong to a path (relative to `src/content` folder). See new feature #91.
+* [New] Added a verbosity level param (optional) to write method of IOInterface.
 * [Fixed] Avoid an error using `--force` option of `new:site` command over a
 non-empty directory that contains "hidden" directories and files.