+++ CHANGELOG.md
@@ -2,9 +2,9 @@ CHANGELOG for 2.x
 =================
 ## 2.2.0
 * [New] Support for themes. See the new feature #98.
-* [New] Package manager with the following new commands: "new:theme", "add:plugin", "remove:plugin" and "update:plugin". Additionaly, "new:site" command has been updated for scaffolding new sites using themes. See the new feature #96.
+* [New] Package manager with the following new commands: `new:theme`, `add:plugin`, `remove:plugin` and `update:plugin`. Additionaly, `new:site` command has been updated for scaffolding new sites using themes. See the new feature #96.
 * [New] Allow default attributes on "pages" collection. See new feature #93.
-* [New] New options for applying "avoid_renderizer" feature to type of files and files belong to a path (relative to `src/content` folder). See new feature #91.
+* [New] New options for applying `avoid_renderizer` feature to type of files and files belong to a path (relative to `src/content` folder). See new feature #91.
 * [New] Added a verbosity level param (optional) to write method of IOInterface.
 * [New] Added verbosity and very-verbosity level messages to the content manager.
 * [New] Added support for overwriting messages to the output in IOInterface.